+++ catcher/UI/Trivial/TrivialUISceneText.swift
@@ -10,12 +10,12 @@ import SpriteKit
 
 class TrivialUISceneText: TrivialUISceneElement {
     let text: String
-    init(position: CGPoint, color: UIColor, delayBeforePresent: Int, text:String) {
+    init(position: CGPoint, color: UIColor, delayBeforePresent: CGFloat, text:String) {
         self.text = text
         super.init(position: position, color: color, delayBeforePresent: delayBeforePresent, name: text)
     }
     
-    convenience init(position: CGPoint, delayBeforePresent: Int, text:String) {
+    convenience init(position: CGPoint, delayBeforePresent: CGFloat, text:String) {
         self.init(position: position, color: UIColor.text(), delayBeforePresent: delayBeforePresent, text: text)
     }
